hw/virtio: Have virtio_bus_get_vdev_bad_features() return 64-bit value
In commit 019a3edbb25 ("virtio: make features 64bit wide") we
increased the 'features' field to 64-bit, but forgot to update
the virtio_bus_get_vdev_bad_features() helper. The 'bad features'
are truncated to 32-bit. The virtio_net_bad_features() handler
from the virtio-net devices is potentially affected.

Have the virtio_bus_get_vdev_bad_features() helper return the
full 64-bit value.

Cc: qemu-stable@nongnu.org
Fixes: 019a3edbb25 ("virtio: make features 64bit wide")
Message-Id: <20210520102822.2471710-1-philmd@redhat.com>

+++ hw/virtio/virtio-bus.c
@@ -134,7 +134,7 @@ size_t virtio_bus_get_vdev_config_len(VirtioBusState *bus)
 }
 
 /* Get bad features of the plugged device. */
-uint32_t virtio_bus_get_vdev_bad_features(VirtioBusState *bus)
+uint64_t virtio_bus_get_vdev_bad_features(VirtioBusState *bus)
 {
     VirtIODevice *vdev = virtio_bus_get_device(bus);
     VirtioDeviceClass *k;
